+++ RoboKit Demo/Testing_RoboKitApp.swift
@@ -5,8 +5,9 @@ import RoboKit
 struct Testing_RoboKitApp: App {
     private var inputSphereManager = InputSphereManager()
     private var formManager: FormManager = FormManager()
-    var client: TCPClient = TCPClient(host: "localhost", port: 12345)
-    @State var socketCollapsed: Bool = true
+    private var client: TCPClient = TCPClient(host: "localhost", port: 12345)
+    private let windowGroupID: String = "WindowGroup"
+    
     @Environment(\.openWindow) private var openWindow
 
     var body: some Scene {
@@ -15,13 +16,13 @@ struct Testing_RoboKitApp: App {
             ImmersiveSpace {
                 ImmersiveView()
                     .onAppear {
-                        openWindow(id: "WindowGroup")
+                        openWindow(id: windowGroupID)
                     }
             }
             .immersionStyle(selection: .constant(.mixed), in: .mixed)
 
-            WindowGroup(id: "WindowGroup") {
-                SocketView(socketCollapsed: $socketCollapsed)
+            WindowGroup(id: windowGroupID) {
+                ControlPanelView()
             }
             .windowStyle(.plain)
             .windowResizability(.contentSize)
